requirements.txt
apiclient==1.0.4
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
Django==4.1.1
django-multiselectfield==0.1.12
djangorestframework==3.13.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.12
